Move Population to seperate file
